refactor(page.tsx): Move the application class to another folder
